created separate file for token api endpoints
