+++ site/app/controllers/AuthenticationController.php
@@ -160,49 +160,6 @@ class AuthenticationController extends AbstractController {
     }
 
     /**
-     * @return MultiResponse
-     */
-    #[Route("/api/token", methods: ["POST"])]
-    public function getToken() {
-        if (!isset($_POST['user_id']) || !isset($_POST['password'])) {
-            $msg = 'Cannot leave user id or password blank';
-            return MultiResponse::JsonOnlyResponse(JsonResponse::getFailResponse($msg));
-        }
-        $this->core->getAuthentication()->setUserId($_POST['user_id']);
-        $this->core->getAuthentication()->setPassword($_POST['password']);
-        $token = $this->core->authenticateJwt();
-        if ($token) {
-            return MultiResponse::JsonOnlyResponse(JsonResponse::getSuccessResponse(['token' => $token]));
-        }
-        else {
-            $msg = "Could not login using that user id or password";
-            return MultiResponse::JsonOnlyResponse(JsonResponse::getFailResponse($msg));
-        }
-    }
-
-    /**
-     *
-     * @return MultiResponse
-     */
-    #[Route("/api/token/invalidate", methods: ["POST"])]
-    public function invalidateToken() {
-        if (!isset($_POST['user_id']) || !isset($_POST['password'])) {
-            $msg = 'Cannot leave user id or password blank';
-            return MultiResponse::JsonOnlyResponse(JsonResponse::getFailResponse($msg));
-        }
-        $this->core->getAuthentication()->setUserId($_POST['user_id']);
-        $this->core->getAuthentication()->setPassword($_POST['password']);
-        $success = $this->core->invalidateJwt();
-        if ($success) {
-            return MultiResponse::JsonOnlyResponse(JsonResponse::getSuccessResponse());
-        }
-        else {
-            $msg = "Could not login using that user id or password";
-            return MultiResponse::JsonOnlyResponse(JsonResponse::getFailResponse($msg));
-        }
-    }
-
-    /**
      * Handle stateless authentication for the VCS endpoints.
      *
      * This endpoint is unique from the other authentication methods in
